deviations.py/run_directory: fix conflict for multiple cores
-After running the file, double check phfile is still there
before trying to delete it. (Used a simple if rather than
catching the exception.)

+++ physics/deviations.py
@@ -159,5 +159,6 @@ def run_directory(filepath, filepattern, savepattern, phpattern, save_subdir='',
         # operate on file
         run_file(f,savefull)
         # remove placeholder
-        os.remove(phfull)
+        if os.path.isfile(phfull):
+            os.remove(phfull)
         print "Finished", f, "at", time.ctime()
